pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hm2p"
version = "0.1.0"
description = "Cloud pipeline for hm2p two-photon calcium imaging in freely-moving mice (RSP, Penk+/CamKII+ populations)"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "Tristan" }]

# ---------------------------------------------------------------------------
# Core runtime dependencies — grouped by pipeline stage
# ---------------------------------------------------------------------------
dependencies = [
    # Configuration & session registry
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",

    # Stage 0 — DAQ / ingest
    "nptdms>=1.9",
    "h5py>=3.11",

    # Stage 3 — Kinematics
    "movement>=0.5",
    "shapely>=2.0",
    "scipy>=1.13",
    "numpy>=1.26",
    "pandas>=2.2",
    "xarray>=2024.6",

    # Stage 4 — Calcium processing
    # NOTE: cascade2p is NOT on PyPI; install separately — see [project.optional-dependencies]
    "roiextractors>=0.5.10",

    # Stage 5 — Sync
    # (numpy + scipy + pandas already listed above)

    # IO & validation
    "pandera>=0.20",
    "pynapple>=0.6",

    # IO — AWS
    "boto3>=1.34",

    # Logging
    "structlog>=24.1",

    # Utilities
    "tqdm>=4.66",
    "rich>=13.7",
    "typer>=0.12",
]

[project.optional-dependencies]
# ---------------------------------------------------------------------------
# Optional — heavy GPU / extraction backends (install in appropriate env)
# ---------------------------------------------------------------------------
suite2p = [
    "suite2p>=0.14",
]
# caiman — NOT listed as an optional dep because the real CaImAn (Flatiron)
# is only on conda-forge, not PyPI (PyPI "caiman" is an unrelated v0.1 package).
# Install via: conda install -c conda-forge caiman
dlc = [
    # DLC 3.0 has only pre-release RCs on PyPI; use 2.3.x stable for now.
    # When 3.0 stable lands, bump to deeplabcut[tf]>=3.0.
    "deeplabcut>=2.3",
]
sleap = [
    "sleap>=1.4",
]
# ---------------------------------------------------------------------------
# Packages below are NOT managed by uv — install manually in separate envs.
# They have incompatible transitive pins (numpy, scikit-learn, tensorflow)
# that break universal resolution. See docs/manual-installs.md.
#
#   fissa>=1.0           — pins scikit-learn<1.2
#   cascade2p            — pins tensorflow==2.3 (Python 3.8 only)
#   caiman (conda-forge) — not on PyPI
#   keypoint-moseq>=0.6  — pins numpy<=1.26
#   vame-py>=0.12        — requires numpy>=2.2
#   cebra>=0.6           — pins numpy<2.0
# ---------------------------------------------------------------------------
nwb = [
    "neuroconv>=0.6",
    "pynwb>=2.7",
]
analysis = [
    "nemos>=0.1",
]
workflow = [
    "snakemake>=8.0",
    "snakemake-executor-plugin-aws-batch>=0.2",
]

# ---------------------------------------------------------------------------
# Dev dependencies
# ---------------------------------------------------------------------------
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "hypothesis>=6.108",
    "ruff>=0.5",
    "mypy>=1.11",
    "bandit[toml]>=1.8",
    "detect-secrets>=1.5",
    "pip-audit>=2.7",
    "vulture>=2.14",
    "pre-commit>=3.8",
    "nbstripout>=0.7",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
]

[project.scripts]
hm2p = "hm2p.cli:app"

# ---------------------------------------------------------------------------
# Tool configuration
# ---------------------------------------------------------------------------

# --- uv --------------------------------------------------------------------
[tool.uv]
# Only resolve for Linux/macOS — this project never runs on Windows.
environments = [
    "sys_platform == 'linux'",
    "sys_platform == 'darwin'",
]
# movement depends on xarray[accel, io, viz]; the viz extra pulls in cartopy
# which needs C libs (libgeos, libproj) we don't need. Override to plain xarray.
override-dependencies = [
    "xarray>=2024.6",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/hm2p"]

# --- ruff -------------------------------------------------------------------
[tool.ruff]
src = ["src"]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "N",   # pep8-naming
    "ANN", # flake8-annotations
    "SIM", # flake8-simplify
]
ignore = [
    "ANN401",  # dynamically typed expressions (Any) — needed for HDF5 attr values
    "SIM108",  # ternary suggestions — explicit if/else more readable for scientific code
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN"]
# F, F0, Fneu are conventional Suite2p / neuroscience variable names —
# suppressing pep8-naming for calcium source + tests + dependent modules.
# T is conventional for number of timepoints in neuroscience.
"src/hm2p/calcium/**"     = ["N803", "N806"]
"src/hm2p/extraction/**"  = ["N803", "N806"]
"tests/calcium/**"        = ["N803", "N806"]
"tests/extraction/**"     = ["N803", "N806"]
"tests/sync/**"           = ["N806"]
"tests/test_cli.py"       = ["N806"]
"src/hm2p/io/hdf5.py"    = ["N806"]
"tests/io/test_hdf5.py"  = ["N803", "N806"]

[tool.ruff.format]
quote-style = "double"

# --- mypy -------------------------------------------------------------------
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["movement.*", "roiextractors.*", "cascade2p.*", "nptdms.*", "pandera.*"]
ignore_missing_imports = true

# --- bandit (security linter) -----------------------------------------------
[tool.bandit]
targets = ["src/"]
exclude_dirs = ["tests/", "old-pipeline/"]
skips = [
    "B101",  # assert — used heavily in tests and validation
    "B404",  # import subprocess — needed for kpms/docker orchestration
    "B603",  # subprocess without shell=True — our use is safe (no user input)
]

# --- pytest -----------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
# Coverage threshold enforced in CI once stages are implemented.
# During skeleton phase, run without --cov-fail-under.
addopts = "--cov=hm2p --cov-report=term-missing"

[tool.coverage.run]
source = ["src/hm2p"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]
